+++ sdk/eventhub/Microsoft.Azure.WebJobs.Extensions.EventHubs/tests/EventHubTests.cs
@@ -7,7 +7,6 @@ using System.Text;
 using Azure.Messaging.EventHubs;
 using Azure.Messaging.EventHubs.Consumer;
 using Azure.Messaging.EventHubs.Primitives;
-using Azure.Storage.Blobs;
 using Microsoft.Azure.WebJobs.EventHubs.Processor;
 using Microsoft.Azure.WebJobs.Host;
 using Microsoft.Azure.WebJobs.Host.TestCommon;
@@ -15,7 +14,6 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
-using Moq;
 using NUnit.Framework;
 
 namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
@@ -84,7 +82,7 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
             Assert.AreEqual(evt.PartitionKey, bindingData["PartitionKey"]);
             Assert.AreEqual(evt.Offset, bindingData["Offset"]);
             Assert.AreEqual(evt.SequenceNumber, bindingData["SequenceNumber"]);
-            Assert.AreEqual(evt.EnqueuedTime, bindingData["EnqueuedTimeUtc"]);
+            Assert.AreEqual(evt.EnqueuedTime.DateTime, bindingData["EnqueuedTimeUtc"]);
             Assert.AreSame(evt.Properties, bindingData["Properties"]);
             IDictionary<string, object> bindingDataSysProps = bindingData["SystemProperties"] as Dictionary<string, object>;
             Assert.NotNull(bindingDataSysProps);
@@ -132,7 +130,7 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
             Assert.AreEqual(events.Length, ((string[])bindingData["PartitionKeyArray"]).Length);
             Assert.AreEqual(events.Length, ((string[])bindingData["OffsetArray"]).Length);
             Assert.AreEqual(events.Length, ((long[])bindingData["SequenceNumberArray"]).Length);
-            Assert.AreEqual(events.Length, ((DateTimeOffset[])bindingData["EnqueuedTimeUtcArray"]).Length);
+            Assert.AreEqual(events.Length, ((DateTime[])bindingData["EnqueuedTimeUtcArray"]).Length);
             Assert.AreEqual(events.Length, ((IDictionary<string, object>[])bindingData["PropertiesArray"]).Length);
             Assert.AreEqual(events.Length, ((IDictionary<string, object>[])bindingData["SystemPropertiesArray"]).Length);
 
@@ -157,29 +155,6 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
             Assert.AreEqual(data, body);
             Assert.Null(contract["PartitionContext"]);
             Assert.Null(contract["partitioncontext"]); // case insensitive
-        }
-
-        // Validate that if connection string has EntityPath, that takes precedence over the parameter.
-        [TestCase("k1", "Endpoint=sb://test89123-ns-x.servicebus.windows.net/;SharedAccessKeyName=ReceiveRule;SharedAccessKey=secretkey")]
-        [TestCase("path2", "Endpoint=sb://test89123-ns-x.servicebus.windows.net/;SharedAccessKeyName=ReceiveRule;SharedAccessKey=secretkey;EntityPath=path2")]
-        public void EntityPathInConnectionString(string expectedPathName, string connectionString)
-        {
-            EventHubOptions options = new EventHubOptions();
-
-            // Test sender
-            options.AddSender("k1", connectionString);
-            var client = options.GetEventHubProducerClient("k1", null);
-            Assert.AreEqual(expectedPathName, client.EventHubName);
-        }
-
-        // Validate that if connection string has EntityPath, that takes precedence over the parameter.
-        [TestCase("k1", "Endpoint=sb://test89123-ns-x.servicebus.windows.net/;SharedAccessKeyName=ReceiveRule;SharedAccessKey=secretkey")]
-        [TestCase("path2", "Endpoint=sb://test89123-ns-x.servicebus.windows.net/;SharedAccessKeyName=ReceiveRule;SharedAccessKey=secretkey;EntityPath=path2")]
-        public void GetEventHubClient_AddsConnection(string expectedPathName, string connectionString)
-        {
-            EventHubOptions options = new EventHubOptions();
-            var client = options.GetEventHubProducerClient("k1", connectionString);
-            Assert.AreEqual(expectedPathName, client.EventHubName);
         }
 
         [TestCase("e", "n1", "n1/e/")]
@@ -246,20 +221,19 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
             // Assert.AreEqual(21, options.PartitionManagerOptions.RenewInterval.TotalSeconds);
         }
 
-        internal static ProcessorPartitionContext GetPartitionContext(string partitionId = "0", string eventHubPath = "path",
+        internal static EventProcessorHostPartition GetPartitionContext(string partitionId = "0", string eventHubPath = "path",
             string consumerGroupName = "group", string owner = null)
         {
-            var processor = new EventProcessorHost.Processor(Int32.MaxValue,
-                consumerGroupName,
+            var processor = new EventProcessorHost(consumerGroupName,
                 "Endpoint=sb://test.servicebus.windows.net/;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=abc123=",
-                "",
                 eventHubPath,
                 new EventProcessorOptions(),
-                null,
-                false,
-                null,
-                Mock.Of<BlobContainerClient>());
-            return new ProcessorPartitionContext(partitionId, processor, s => default);
+                Int32.MaxValue,
+                false, null);
+            return new EventProcessorHostPartition(partitionId)
+            {
+                ProcessorHost = processor
+            };
         }
     }
 }